requirements.txt
duckdb==0.9.2
matplotlib==3.8.2
numpy==1.26.3
pandas==2.1.4
Pillow==10.2.0
plotly==5.18.0
seaborn==0.13.1
streamlit==1.29.0
